asr.py
import os
import json
from groq import Groq
from dotenv import load_dotenv

load_dotenv()

GROQ_API_KEY = os.getenv('GROQ_API_KEY')

def transcribe_audio(audio_file_path):
    client = Groq(api_key=GROQ_API_KEY)
    try:
        with open(audio_file_path, "rb") as file:
            transcription = client.audio.transcriptions.create(
                file=file,
                model="whisper-large-v3-turbo",
                response_format="verbose_json",
                timestamp_granularities=["word", "segment"],
                language="en",
                temperature=0.0
            )
            return transcription
    except Exception as e:
        raise Exception(f"Error during transcription: {str(e)}")